Add kubelet-config.yaml parser.
Add parser for kubelet-config.yaml to set the
"kubeReserved.cpu" field. The kubeReserved field along with
the available CPUs on a machine sets the Allocatable CPUs field, which
is considered when scheduling Pods in the cluster
(Allocated CPU = CPUs - kubeReserved - evictionThreashold). This change
adds the logic to compute the value of kubeReserved based on
GKE's formula for computing reserved cpu: https://cloud.google.com/kubernetes-engine/docs/concepts/cluster-architecture#memory_cpu

PiperOrigin-RevId: 352132084